ajout des autorisations pour la function deleteOneAdvisor du controller advisors.js

+++ Controllers/advisors.js
@@ -1,4 +1,23 @@
 const Advisor = require("../Models/Advisor");
+const Manager = require('../Models/Manager');
+const Director = require('../Models/Director');
+const mongoose = require('mongoose');
+
+
+const checkManagerId = (manager_id, advisor_id) => {
+    return Advisor.findOne({ managerId: manager_id, _id: advisor_id })
+    .then((advisor) => {
+        if(advisor) return true;
+        else return false;
+    })
+}
+const checkDirectorId = async (director_id) => {
+    return Director.find({ _id: director_id })
+    .then((director) => {
+        if(director) return true;
+        else return false;
+    })
+}
 
 const createAdvisor = (req, res) => {
     const newAdvisor = new Advisor({
@@ -55,15 +74,46 @@ const putAdvisor = (req, res) => {
     })
 }
 
-const deleteOneAdvisor = (req, res) => {
+const deleteOneAdvisor = async (req, res) => {
     const id = req.params.id;
-    return Advisor.findByIdAndDelete(id)
-    .then((advisorToDelete) => {
-        return res.send(`Advisor ${advisorToDelete.firstname} ${advisorToDelete.lastname} has been deleted !`);
-    })
-    .catch((error) => {
-        return res.status(400).send(error);
-    })
+    const userId = req.headers.userid;
+    const isValidUser = mongoose.isValidObjectId(userId);
+    const isValidAdvisor = mongoose.isValidObjectId(id);
+    if (isValidAdvisor === false) {
+        return res.status(400).send("l'id de l'advisor n'est pas valide");
+    }
+    const advisor = await Advisor.findById(id);
+    if (isValidUser === true) {
+        const manager = await Manager.findById(userId);
+        const director = await Director.findById(userId);
+        let checkManager;
+        let checkDirector;
+        if (manager != null || manager != undefined) {
+            try{
+                checkManager = await checkManagerId(manager._id, advisor);
+            } catch (error) {
+                return res.status(500).send(error);
+            }
+        }
+        if (director != null || director != undefined) {
+            try {
+                checkDirector = await checkDirectorId(director._id);
+            } catch (error) {
+                return res.status(500).send(error);
+            }
+        }
+        if (checkManager === true || checkDirector === true && isValidUser === true) {
+            return Advisor.findByIdAndDelete(id)
+            .then(() => {
+                return res.send(`Advisor ${advisor.firstname} ${advisor.lastname} has been deleted !`);
+            })
+            .catch((error) => {
+                return res.status(400).send(error);
+            })
+        } else {
+            return res.status(403).send('You don\'t have access to this advisor\'s data !');
+        }
+    }
 }
 module.exports = {
     createAdvisor: createAdvisor,
